Refactor deploy.WithJoinOptions to join.Options
