+++ pkg/join/options.go
@@ -1,10 +1,14 @@
 /*
 SPDX-License-Identifier: Apache-2.0
+
 Copyright Contributors to the Submariner project.
+
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
+
     http://www.apache.org/licenses/LICENSE-2.0
+
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -12,9 +16,9 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package deploy
+package join
 
-type WithJoinOptions struct {
+type Options struct {
 	PreferredServer               bool
 	ForceUDPEncaps                bool
 	NATTraversal                  bool
